the last commit

+++ src/main/java/com/nam/libraryapplication/services/StudentService.java
@@ -1,6 +1,9 @@
 package com.nam.libraryapplication.services;
 
+import com.nam.libraryapplication.entities.Book;
+import com.nam.libraryapplication.entities.Renting;
 import com.nam.libraryapplication.entities.Student;
+import com.nam.libraryapplication.repos.RentingRepo;
 import com.nam.libraryapplication.repos.StudentRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Example;
@@ -14,6 +17,9 @@ import java.util.List;
 public class StudentService {
     @Autowired
     StudentRepo studentRepo;
+
+    @Autowired
+    RentingRepo rentingRepo;
     public ResponseEntity<String> createStudent(Student student) {
         try {
             Example<Student> studentExample = Example.of(student);
@@ -46,11 +52,43 @@ public class StudentService {
     }
 
     public ResponseEntity<String> deleteStudentById(Integer id) {
-        if(studentRepo.existsById(id)) {
-            studentRepo.deleteById(id);
-            return new ResponseEntity<>("Delete Success", HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>("Could not find the following ID !!!\nDeleted Failed !!!", HttpStatus.BAD_REQUEST);
+        try {
+            if(studentRepo.existsById(id)) {
+                Student studentToDelete = studentRepo.findById(id).orElse(null);
+                if(studentToDelete != null) {
+                    List<Renting> rentsContainingStudent = rentingRepo.findAllByStudentID(studentToDelete);
+                    if(!rentsContainingStudent.isEmpty()) {
+                        return new ResponseEntity<>("You must delete rents that contain this student first !!!", HttpStatus.BAD_REQUEST);
+                    } else{
+                        studentRepo.deleteById(id);
+                        return new ResponseEntity<>("Deleted Success !!!", HttpStatus.OK);
+                    }
+                }
+            }
+        } catch(Exception e) {
+            e.getStackTrace();
         }
+        return new ResponseEntity<>("Could not find the following ID !!!\nDeleted Failed !!!", HttpStatus.BAD_REQUEST);
     }
+
+
+//    public ResponseEntity<String> deleteBookById(Integer id) {
+//        try {
+//            if(bookRepo.existsById(id)) {
+//                Book bookToDelete = bookRepo.findById(id).orElse(null);
+//                if(bookToDelete != null) {
+//                    List<Renting> rentsContainingBook = rentingRepo.findByBookIDContaining(bookToDelete);
+//                    if(!rentsContainingBook.isEmpty()) {
+//                        return new ResponseEntity<>("You must delete rents that contain this book first !!!", HttpStatus.BAD_REQUEST);
+//                    } else{
+//                        bookRepo.deleteById(id);
+//                        return new ResponseEntity<>("Deleted Success !!!", HttpStatus.OK);
+//                    }
+//                }
+//            }
+//        } catch(Exception e) {
+//            e.getStackTrace();
+//        }
+//        return new ResponseEntity<>("Could not find the following ID !!!\nDeleted Failed !!!", HttpStatus.BAD_REQUEST);
+//    }
 }
